requirements.txt
wget==3.2
psycopg2-binary==2.9.3